[fix] frame number int to double

+++ src/psnr.cpp
@@ -4,11 +4,11 @@
 #define pixel_range 255
 
 void mse(unsigned char *in1, unsigned char *in2, int *resoloution, double *out_sqrt_mse){
-	int tmp_mse = 0;
+	double tmp_mse = 0.0;
 	for( int i=0 ; i < *resoloution ; i++ ){
 		tmp_mse += (in1[i] - in2[i]) * (in1[i] - in2[i]); 
 	}
-	*out_sqrt_mse = sqrt((double)tmp_mse / (double)*resoloution);
+	*out_sqrt_mse = sqrt(tmp_mse / (double)*resoloution);
 }
 
 void psnr(double *in_sqrt_mse, double *out_psnr){
